+++ source/Application/Model/ArticleList.php
@@ -1084,8 +1084,8 @@ class ArticleList extends \OxidEsales\Eshop\Core\Model\ListModel
         $sSelectFields = $oBaseObject->getSelectFields();
 
         $sSelect = "select {$sSelectFields} from {$sArticleTable} where oxvarminprice >= 0 ";
-        $sSelect .= $dPriceTo ? "and oxvarminprice <= " . (double) $dPriceTo . " " : " ";
-        $sSelect .= $dPriceFrom ? "and oxvarminprice  >= " . (double) $dPriceFrom . " " : " ";
+        $sSelect .= $dPriceTo ? "and oxvarminprice <= " . (float) $dPriceTo . " " : " ";
+        $sSelect .= $dPriceFrom ? "and oxvarminprice  >= " . (float) $dPriceFrom . " " : " ";
 
         $sSelect .= " and " . $oBaseObject->getSqlActiveSnippet() . " and {$sArticleTable}.oxissearch = 1";
