+++ source/Application/Model/Order.php
@@ -1656,7 +1656,7 @@ class Order extends \OxidEsales\Eshop\Core\Model\BaseModel
         ];
 
         // We force reading from master to prevent issues with slow replications or open transactions (see ESDEV-3804).
-        return (double) \OxidEsales\Eshop\Core\DatabaseProvider::getMaster()->getOne($sSelect, $params);
+        return (float) \OxidEsales\Eshop\Core\DatabaseProvider::getMaster()->getOne($sSelect, $params);
     }
 
     /**
@@ -1904,7 +1904,7 @@ class Order extends \OxidEsales\Eshop\Core\Model\BaseModel
     {
         $oCur = \OxidEsales\Eshop\Core\Registry::getConfig()->getActShopCurrencyObject();
 
-        return number_format((double) $this->oxorder__oxtotalordersum->value, $oCur->decimal, '.', '');
+        return number_format((float) $this->oxorder__oxtotalordersum->value, $oCur->decimal, '.', '');
     }
 
     /**
@@ -1918,10 +1918,10 @@ class Order extends \OxidEsales\Eshop\Core\Model\BaseModel
     {
         $aVats = [];
         if ($this->oxorder__oxartvat1->value) {
-            $aVats[$this->oxorder__oxartvat1->value] = $this->oxorder__oxartvatprice1->value;
+            $aVats[(int) $this->oxorder__oxartvat1->value] = $this->oxorder__oxartvatprice1->value;
         }
         if ($this->oxorder__oxartvat2->value) {
-            $aVats[$this->oxorder__oxartvat2->value] = $this->oxorder__oxartvatprice2->value;
+            $aVats[(int) $this->oxorder__oxartvat2->value] = $this->oxorder__oxartvatprice2->value;
         }
 
         if ($blFormatCurrency) {
